+++ doc/changelog/latest.rst
@@ -42,6 +42,7 @@ Changed
   :meth:`~tenpy.models.mixed_xk.MixedXKModel.add_intra_ring_hopping`,
   :meth:`~tenpy.models.mixed_xk.MixedXKModel.add_inter_ring_interaction`, and
   :meth:`~tenpy.models.mixed_xk.MixedXKModel.add_intre_ring_interaction` to vary with `x`.
+- Renamed the module :mod:`~tenpy.linalg.lanczos` to :mod:`tenpy.linalg.krylov_based`.
 
 Fixed
 ^^^^^
