+++ tenpy/networks/mps.py
@@ -158,7 +158,7 @@ logger = logging.getLogger(__name__)
 
 from ..linalg import np_conserved as npc
 from ..linalg import sparse
-from ..linalg.lanczos import Arnoldi
+from ..linalg.krylov_based import Arnoldi
 from .site import GroupedSite, group_sites
 from ..tools.misc import to_iterable, to_array, get_recursive
 from ..tools.math import lcm, speigs, entropy
